Add no results handling

+++ src/components/Recipe.js
@@ -1,8 +1,9 @@
 import '../styles/Recipe.css';
 import React, { useState } from 'react';
 
-function Recipe({ description, id, imageUrl, isLiked, link, steps, title, addRecipe, removeRecipe }) {
+function Recipe({ description, id, imageUrl, isLiked = false, link, steps, title, addRecipe, removeRecipe }) {
   const [ liked, setLiked ] = useState(isLiked);
+console.log(isLiked);
 
   function like() {
     if(liked === false) {
@@ -15,7 +16,7 @@ function Recipe({ description, id, imageUrl, isLiked, link, steps, title, addRec
       })
       .then(res => res.json())
       .then(data => addRecipe(data));
-    } else {
+    } else if(liked === true) {
       fetch(`http://localhost:3000/recipes/${id}`, {
         method: 'DELETE',
         headers: {
